+++ cellmonitor/source/android/CellMonitor/app/proguard-rules.pro
