+++ cellmonitor/source/android/CellMonitor/app/src/main/AndroidManifest.xml
