+++ cellmonitor/source/android/CellMonitor/app/src/main/java/org/kaaproject/kaa/demo/cellmonitor/util/LocationUtil.java
