+++ cellmonitor/source/android/CellMonitor/app/src/main/res/drawable-mdpi/ic_launcher.png
