+++ cellmonitor/source/android/CellMonitor/app/src/main/res/layout/activity_cell_monitor.xml
