+++ cellmonitor/source/android/CellMonitor/app/src/main/res/values-v14/styles.xml
