+++ cellmonitor/source/android/CellMonitor/app/src/main/res/values-w820dp/dimens.xml
