+++ cellmonitor/source/android/CellMonitor/app/src/main/res/values/strings.xml
