+++ cellmonitor/source/android/CellMonitor/app/src/main/res/values/styles.xml
