+++ cellmonitor/source/android/CellMonitor/gradle.properties
