--- cellmonitor/source/android/app/src/androidTest/java/org/kaaproject/kaa/demo/cellmonitor/ApplicationTest.java
@@ -1,13 +1,0 @@
-package org.kaaproject.kaa.demo.cellmonitor;
-
-import android.app.Application;
-import android.test.ApplicationTestCase;
-
-/**
- * <a href="http://d.android.com/tools/testing/testing_android.html">Testing Fundamentals</a>
- */
-public class ApplicationTest extends ApplicationTestCase<Application> {
-    public ApplicationTest() {
-        super(Application.class);
-    }
-}
