--- cellmonitor/source/android/app/src/main/res/values/colors.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <color name="red">#FF0000</color>
-</resources>
