--- cellmonitor/source/android/app/src/test/java/org/kaaproject/kaa/demo/cellmonitor/ExampleUnitTest.java
@@ -1,15 +1,0 @@
-package org.kaaproject.kaa.demo.cellmonitor;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-/**
- * To work on unit tests, switch the Test Artifact in the Build Variants view.
- */
-public class ExampleUnitTest {
-    @Test
-    public void addition_isCorrect() throws Exception {
-        assertEquals(4, 2 + 2);
-    }
-}
